move functions into users
